requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
Flask-JWT-Extended
Flask-SocketIO
eventlet
marshmallow
flask-cors 
Flask-SQLalchemy